tast: Set crash test in progress during crash tests
The autotest framework surrounding tast will touch
/mnt/stateful_partition/etc/collect_chrome_crashes while running to
tell crash_reporter to grab more data during crashes. This is very
helpful for debugging non-crash tast tests, but is problematic for
actual crash_reporter tests. In particular, it tells the UserCollector
to write out crash reports for Chrome; these can collide with crash
reports being generated by the ChromeCollector, causing one of the
two crash_reporters to fail out unexpectedly.

I'm certainly seeing this in some of the ChromeCrashLoggedIn flakes
(see /var/log/messages in
https://stainless.corp.google.com/browse/chromeos-autotest-results/swarming-46c42d8fe0bbb310/autoserv_test/tast/results/tests/ui.ChromeCrashLoggedIn/).
I'm not sure it's the actual *cause* of the flakes, but if it's not,
it's causing enough error messages that I can't find the real cause.
In any cause, telling crash_reporter to act more like the production
version (not use UserCollector for Chrome crashes) makes the tests more
accurate.

So, during crash_reporter tests, touch
/run/crash_reporter/crash-test-in-progress to move the behavior back
closer to normal behavior.

BUG=chromium:984807
TEST=tast -verbose run --extrauseflags chrome_internal my_crbook ui.ChromeCrash*

Change-Id: Id8a2344a3bfc2fab3dedb4c50e7c6e56bcca4aca