Update patch set 4
Patch Set 4:

(2 comments)

My main concern with a test like this is that it will be fragile to the UI changing something like the ID of an element that's under test here. Since this test won't catch it until the changes have already rolled into Chrome, then we'll start getting spurious failures after the fact (and those might in turn hide real failures if something else is going on).

Can we use the private autotest API for this?

Patch-set: 4
Reviewer: Gerrit User 1117055 <1117055@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>