video.DecodeAccelPerf: Don't use --gtest_repeat=-1 in performance tests.
To measure video decoder CPU usage a video is played in real-time while CPU
usage is measured, using the video_decode_accelerator_perf_tests. As the video
might be shorter than the measurement duration we use --gtest_repeat=-1 to
repeat video playback. At the end of the measurement interval the video decoder
test process is killed.

Unfortunately this can cause some unexpected issues if the --gtest_filter that
is used matches no tests. This can happen e.g. when a test was renamed or the
wrong filter was used (e.g. b/138822793). In this case the test run skips all
tests and repeats this indefinitely, spamming the command line with output:
> [==========] Running 0 tests from 0 test suites.
> [==========] 0 tests from 0 test suites ran. (0 ms total)
> [  PASSED  ] 0 tests.
> Repeating all tests (iteration 897894) . . .

This generates huge log files (e.g. 800+MB for just 20s), causing various
hard-to-debug issues due to the disk being full.

Rather than repeating indefinitely we use a sufficiently high value, so the test
is guaranteed to run for the entire measurement duration. The performance tests
currently play a 10s video and measure cpu usage for about 20-30s, so repeating
about 5 times should be more than sufficient. We use '1000' here, to make sure
the test will also repeat long enough if we choose to use shorter videos in the
future, but small enough that the log size would be acceptable on filter
mismatch (about 200kb).

TEST=video.DecodeAccelPerfH2641080P60FPS on eve

BUG=b:138822793

Cq-Depend: chromium:1740150
Change-Id: I4d40c3467873b23a302c6ada1b2a4a5c61929bae
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/1745927
Commit-Queue: David Staessens <dstaessens@chromium.org>
Tested-by: David Staessens <dstaessens@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>