arc.MiniContainerState arc.ContainerMount: use effective_cpus instead of cpus
On upstream Linux cgroup cpu controller the effective_cpus is
identical to cpus.

Kernels with
https://android-review.googlesource.com/c/kernel/common/+/270102
changes this and aligns more closely to the cgroups v2 behavior. In this
case effective_cpus shows the upstream expected values and cpus returns
the last configured value even if some of those cpus are currently
offline.

Change the tests to always use effective_cpus which is consistent with or
without the above kernel patch.

BUG=b:136684390
TEST=tast run {} arc.MiniContainerState arc.ContainerMount

Change-Id: I2c6688edc1665d7c080b3700f6f6f39c953cede3