tast-tests: Set precise timeout for precondition Prepare.
This CL sets timeout for each step in Prepare() for ARC and Chrome
preconditions.
Specifically, this introduces the timeout of "reset" step.
It turend out that in some cases a previous test makes the testing
environment broken somehow so that reset step timed out, so that
whole Prepare() was timed out.
With this CL, on timeout of reset step, it tries to restart the
precondition from scratch.

Also, this CL adds timing measurement for the reset step.

BUG=chromium:992447
TEST=Ran locally.

Change-Id: I43d1d1a021a22ef43d63d7be49d899e33075b7de