webrtc: Move WebRTC PeerConn tests to a new package 'webrtc'
This CL is the 1st CL of 3 CLs to move WebRTC-related tests to 'webrtc' package.

There has been camera and video tests for webrtc use case. The code for the
tests are put in camera and video directory. We should put them in webrtc
directory so that it represents where bugs might be.

This CL creates a new directory, "webrtc", and moves WebRTC PeerConn tests to the
directory.
Also, this renames camera.WebRTC(Perf) test to camera.GetUserMedia as they don't
exercise WebRTC stack.

* camera.WebRTCPeerConn* -> webrtc.PeerConn*
* camera.WebRTC -> camera.GetUserMedia
* camera.WebRTCPerf -> camera.GetUserMediaPerf

BUG=chromium:971922
TEST=Ran tast.webrtc.* on kevin
TEST=Ran tast.camera.GetUserMedia* on kevin
TEST=Ran tast.video.WebRTC* on kevin

Change-Id: I630f414be00a3845ad441538d5ec3bcfef2b558f
Signed-off-by: Keiichi Watanabe <keiichiw@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/1769960
Reviewed-by: Shuhei Takahashi <nya@chromium.org>