PopAllUserProfiles() before calling ConfigureServiceForProfile()
In light of recent Ethernet changes (crosreview.com/1763916) and
because ethernet_any is now in many user profiles, running
ConfigureServiceForProfile() with the profile argument set to the
default profile path causes a check to fail (the profile of the
service to be configured must be equal to or before the profile
argument passed in) and an error ('Temporary service configured
but not usable') to be returned.

Call PopAllUserProfiles() so that the ethernet_any service's profile
will be the default profile.

BUG=None
TEST=test passes

Cq-Depend: chromium:1763617
Change-Id: I1a01236dda058dd8f54105519e3223ec7094872b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/1772324
Tested-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Matthew Wang <matthewmwang@chromium.org>