Fix arc.VirtualDesks test
The test used to assume that a browser window is created by default
and used to test that it remained on desk one. Apprently, this
recently changed and a browser window no longer starts automatically.

This CL makes the test more deterministic by explicitly starting
a new browser window at the beginning of the test.

BUG=b:140218934
TEST=tast -verbose run $DUT arc.VirtualDesks

Change-Id: If36229d9c58c7c8af3947abf1e20b7b296eb2567