Add Fake DMServer startup/cleanup to Tast.
Add a struct and various methods to set up, start, and kill policy_testserver.py.

Major pieces:
- FakeDMS struct to be used by the caller
- New() method to set up server args
- start() method to start the command and wait for either a successful ping or
for the command to fail
- Stop() method to end the process gracefully or kill it if needed

A note on ports:
The Autotest version of this code gets an unused port the same way I have done
here.  policy_testserver.py is using a basic python http server, so
theoretically I could set the port to 0 and have the os choose one.  But
there's currently no great way to get that port from the server.  If this code
ends up being flaky (which it is not in Autotest), then I could either modify
policy_testserver or use netstat or some such tool.

BUG=chromium:978654
TEST=ran on an DUT with the policy-testserver ebuild installed, saw that
FakeDMS is able to start, set policies (a functionality not included in this
CL), and stop.
Cq-Depend: chromium:1661334
Change-Id: Id55e2843916028d185a61cf0fe1d6a6c081f0a96