Update patch set 6
Patch Set 6: Code-Review+2

OK. The constants seem fine. Thanks.

I'm definitely less of an expert on the typing, so it's not clear to me about what the advantages/disadvantages are for defining a type as an alias to 'string'. I'll leave that to Ricardo, who suggested it in the first place. (And maybe I'll read the spec a bit more when I get some time...)

Patch-set: 6
Reviewer: Gerrit User 1132414 <1132414@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+2