tast-tests: Add usbc/fluffy_interop_std.go.
This commit adds the first tast test using fluffy.  It is a basic
interoperability test and the config is specified in
fluffy_interop_std_config.json.

It simply connects each charger to the DUT and measures the voltage at
the DUT and verify that the voltage is expected.

BUG=b:140451137
TEST=`tast -verbose run -checktestdeps=false -var=usbc.MaxPwrReqMW=45000
<DUT IP> usbc.FluffyInteropStd`

Change-Id: I39b7fcdea33569dd5a2290902c542e8019346658
Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/1779614
Tested-by: Aseda Aboagye <aaboagye@chromium.org>
Commit-Queue: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Auto-Submit: Aseda Aboagye <aaboagye@chromium.org>