camera.CCA*: Simplify and fix the app launching flow
The background page might be killed by Chrome when idle. Use
onMessageExternal listener to prevent race condition when running test.

The companion CL in CCA is CL:1781924. Because it's a fundamental change
of app startup flow and there is easy way to make it backward
compatible, we will wait the CCA CL land and pass Chrome PFQ first, and
then land this CL.

BUG=b:140358697
TEST=With a manually added sleep(20), do
     $ tast run <dut> camera.CCAUIPreviewPerf

Change-Id: Id42a73f10767d2c1fa3b394b668cf9519a4634f8