video: Explore metrics file in under the output directory instead of hard-code
The present tast code assumes the performance metrics are saved in the output
directory in video_decode_accelerator_perf_tests. crrev.com/c/1778104 is going
to change the metrics file place to the sub directory. For the chromium CL, this
CL removes the assumption by exploring the metrics file under the output
directory.

Bug: None
Test: tast run video.DecodeAccelVDPerfH2641080P30FPS
Change-Id: I8135794887a52681b7c61d8d588c4af999aa4c64