video: Add a processing for a new metrics file path in decode perf tests
The present tast code assumes the performance metrics are saved in the output
directory in video_decode_accelerator_perf_tests. crrev.com/c/1778104 is going
to change the metrics file place to the sub directory. Concrectely, it is
changed from tmpOutDir/MeasureCappedPerformance.json to tmpOutDir/VideoDecoderTest/
MeasureCappedPerformance.json. So that the test will not fail due to the change,
this CL changes the tast code to check both the new and old file places. The old
place check should be removed once crrev.com/c/1778104 is landed.

BUG=chromium:917951
TEST=tast run video.DecodeAccelVDPerfH2641080P30FPS w/wo crrev.com/c/1778104
Change-Id: I8135794887a52681b7c61d8d588c4af999aa4c64