tast-test: ash/wm support PIP window state
This CL adds the Picture-in-Picture window state to WindowStateType.
This is needed to replace the existing "dumpsys | regexp" with a more
robust way to query for PIP window state.

Addtionally this CL also fixes some nits in the documentation.

(Depends on https://crrev.com/c/1785058, but can't use it in Cq-Depends
since it is outside the scope)

TEST=Called "ash.GetARCAppWindowState()" on an ARC window and it
     returned "PIP"
BUG=chromium:1000774

Change-Id: I7e53f026144738c53bc300fc764abe62d5b018ed