Update patch set 2
Patch Set 2:

(4 comments)

Thank you for offline discussion.

Could you check the compiler options and linker options?
Specifically, it looks important of whether or not the functions in bomb is inlined or not.
Also, UB in recbomb() needs to be double-checked if it can produce a binary needed for the testing.

Patch-set: 2
Reviewer: Gerrit User 1115919 <1115919@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>