tast-tests: Add tooltip property test
https://crrev.com/c/1741328 Adds the tooltip property (in preparation
for adding tooltip actions) to automation nodes.
This CL also adds a node struct to accessibility.go
The intention of adding the struct is to reduce number of parameters
which gets sent to the focusAndCheckElement function.
This should make it easier and simpler to add more properties in
tast tests for accessibility.

CL which adds tooltips to elements (ag/9235164).

TEST=tast run DUT_IP arc.accessibilityEvent
BUG=b:137246277

Change-Id: I23631678d4f7985483af514f81524105453112a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/1772904
Commit-Queue: Sara Kato <sarakato@chromium.org>
Tested-by: Sara Kato <sarakato@chromium.org>
Auto-Submit: Sara Kato <sarakato@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>