video.DecodeAccelPerf: Fail test if test binary crashes while measuring cpu.
When measuring CPU usage while running a process, the tests currently don't
check whether the test binary executed succesfully. This means that if the test
binary crashes while measuring CPU usage, the invalid CPU usage results are
still reported.

This caused b/140534129 to go undetected for a while. This CL makes sure the
tests fail if a process crashed.

TEST=video.DecodeAccelPerfH2641080P30FPS on eve

BUG=b:140533227

Change-Id: Icda2adac5839518b7a8a3dc53417bcfa46d38e67