video.PlaybackPerf*: Increase timeout of plackback performance tests.
This CL increases the timeout of all playback performance tests. Even on
successful runs many devices seem to take 2m45+ for the test to complete, which
is much too close to the configured limit of 3 minutes. This causes the playback
performance test to be flaky on multiple boards.

The timeout is increased from 3 minutes to 5 minutes to leave a bit of margin to
prevent future flakiness.

TEST=video.video.PlaybackPerfVP92160P60FPS on eve

BUG=chromium:989370

Change-Id: Ifecb028293fffb82201a34bd89f7cef21caf8337
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/1786461
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Tested-by: David Staessens <dstaessens@chromium.org>
Commit-Queue: David Staessens <dstaessens@chromium.org>