Create exceptions for security check for moblab's use of docker.
Moblab has started to use docker containers to run the envoy proxy.
This enables the new UI which uses grpc backends and grpc-web from
the client to communicate to them, which requires the envoy proxy.

https://ci.chromium.org/p/chromeos/builders/testplatform/cros_test_platform/b8902819824923681184
shows a clean test run.

( The tryjob shows failures https://ci.chromium.org/p/chromeos/builders/general/Try/b8902828310621192256
but that is due to an infra issue )

BUG=None
TEST=tryjob

Change-Id: I2b76dc182bc20055ed7f667200960a2b6ea50907