tast-tests: Support faillog screencapture with Chrome API.
This CL starts to try screencapture by using Chrome API.
If failed, it falls back to current screenshot method, which is based
on screenshot command line.

The failure happens in various situation, specifically, if Chrome does not
listening the debug port.
We could not rescue such cases by this change, but this CL still improves
the coverage.

BUG=chromium:984390
TEST=Ran locally.

Change-Id: Ibbb4aecfa08c263b48c387711f00a70badd44ef1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/1792383
Tested-by: Hidehiko Abe <hidehiko@chromium.org>
Commit-Queue: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>