usbc.fluffyInteropStd: Add variables for servod connection spec
This commit adds the ability for the usbc.fluffyInteropStd test to take
in two runtime variables that define the servod hostname and port.  The
variables are "ServodHost" and "ServodPort".

BUG=none
TEST=`tast -verbose run -checktestdeps=false -continueafterfailure=true
-var=usbc.MaxPwrReqMW=45000 -var=ServodPort=9998 100.90.31.239
usbc.FluffyInteropStd`; verify that tast attempts to connect to servod
at localhost on port 9998.
TEST=Set `-var=ServodHost=100.90.31.2`, verify that tast attempts to
connect to servod at that IP with port 9999.

Change-Id: I458d231cd22de53305190d8f9b6a3b6be51d2d7e
Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/1794107
Tested-by: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-by: Mengqi Guo <mqg@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Aseda Aboagye <aaboagye@chromium.org>
Auto-Submit: Aseda Aboagye <aaboagye@chromium.org>