tast: Adds ARC multicast forwarder test
This adds testing for multicast forwarding in ARC. It checks
if supported multicast protocols mDNS, legacy mDNS, and SSDP
are forwarded correctly.

There are outbound and inbound forwarding check.

For outbound, this is done by asserting that packets sent by
ARC will be forwarded and will be sent by Chrome OS.

For inbound, multicast messages are sent with loopback flag
turned on. The test will assert that the packets are received
by ARC.

BUG=b:136710799
TEST=tast run <DUT> arc.MulticastForwarder (on nocturne)

Cq-Depend: chromium:1786525, chromium:2035451
Change-Id: Ibf4a76bab825891875f2b66fc8cb67e89341aec7