tast-test: activity.WindowBounds return bounds correctly
It seems that on certain conditions freeform activities can include the
caption height inside their bounds.

When this condition happens we should not add the caption height in
their bounds since it is already part of the bounds.

The condition when this occurs is when bounds.top == 0.

This CL addresses that condition.

TEST=arc.SetBounds(), arc.SurfaceInsets(), arc.PIP work as expected.
BUG=chromium:1003116

Cq-Depnd: chromium:1799363
Change-Id: I7d9ce537cb5f13181e8a49b3b23b982ceaba3af8