video.DecodeAccel*: Store crash dumps when video decoder tests crash.
This CL changes the video.DecodeAccel* tests to copy any crash dump generated
by the video_decode_accelerator_tests binary to the test directory, so they will
be stored as test artifacts.

These crash dumps will be very helpful when debugging test crashes. The size of
the crash dumps can run into the tens of megabytes. But as they are only
collected for our binary video decoder tests in case of test crashes, the impact
should be limited.

TEST=video.DecodeAccelH264New on eve

BUG=b:141656083

Change-Id: I90884b1cbd5aadf7ee2af76b98aa5a56edb18ca0