video.DecodeAccel*: Store stack traces when video decoder tests crash.
This CL changes the video.DecodeAccel* tests to store stack traces generated
by the video_decode_accelerator_tests binary as test artifacts. These stack
traces will be very helpful when debugging test crashes.

The size of crash dumps can run into the tens of megabytes. As the crash dumps
themselves are too large to be stored by the test infrastructure, we extract the
stack trace from the core dump files using gdb.

These stack traces are a few kB and are only collected for our binary video
decoder tests in case of test crashes, so the impact on the total size of test
artifacts should be limited.

TEST=video.DecodeAccelH264New on eve

BUG=b:141656083

Change-Id: I90884b1cbd5aadf7ee2af76b98aa5a56edb18ca0