Add webView element for accessibility test.
Updated apk is available here (ag/9342087).

Previously in N, enabling ChromeVox on play store resulted in a crash.
This crash occurs due to ChromeVox trying to focus on a webView element.

This test verifies the fix (crrev/1730654), by ensuring that ChromeVox does
not crash on applications that have a webView element. The test enables
ChromeVox on such an application, and ensuring that the crash does not
occur.

BUG=b:136621529
TEST=tast run DUT_IP arc.AccessibilityEvent/AccessibilityTree
Change-Id: I109a5a83f83b398bb183cf4d92fdcd1c7495c2d8