Update patch set 1
Patch Set 1:

I'm not convinced about this solution. Could you add more info why this is needed? Thanks.

there is a bug filed for this: https://crbug.com/1002958
I think we should fix this bug instead.

and the reason I'm not convinced is if we follow this pattern, we will need a rotateScreen.RunTest() (assuming that there is a bug in rotate screen)... and then what if we need to combine tablet mode with screen rotation? We will need a rotateScreenInTabletMode.runTest()

Patch-set: 1
CC: Gerrit User 1202079 <1202079@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>