tast: Give crash tests a clean crash directory.
If /var/spool/crash is full when a crash test starts, it will have
unexpected failures due to being unable to save crashes. Instead, start
it with an empty directory.  (It's still possible this will flake if
another process is _very_ crashy to the point that it crashes 32 times
during the test run, but that's relatively unlikely. If it becomes a
problem, we can also edit core_pattern to use --filter_in to ignore
other crashes during the test run.)

TEST=tast -verbose run $DEVICE_IP platform.KernelWarning on a device with a full spool directory
BUG=chromium:998986

Change-Id: Idb1bd366d3573efde4d4e271b4ade54ad1adba61