tast-tests: Add sleep for accessibilityEvent
Ongoing flakiness has been observed with accessibilityEvent.
(b/123397142)

This CL adds a "sleep" after ChromeVox has been activated, and
before interacting with the UI. The proposal is to add sleep,
and see if the flakiness of the test improves. This will aid in
determining whether or not changing the serialization/ sampling
timing of the accessibility tree will help to address many of the
issues for ARC++ accessibility (b/131123212).

This is temporary change, and the sleep will be removed after
some time.

Bug: b/131123212
Bug: b/123397142
Test: arc.accessibilityEvent is flaky
Change-Id: If0d2f74c95a424d4a0531339f9f1610499089f0b