tast_tests: crostini: Increase timeout on RunWindowedApp
In tablet mode requesting a keyboard device creates a one off virtual
keyboard, which waits 5 seconds to give Chrome a chance to notice the
new device. This causes the crostini copy paste tests fail
consistently in tablet mode because we only give RunWindowedApp a 5
second timeout. Fix this by bumping up the timeout to 10 seconds.

BUG=chromium:997564
TEST=tast -verbose run 100.124.162.50 crostini.CopyPasteArtifact*

Change-Id: I4d5a1d119d84d823d940d9c2a4c8f7a0ea9b99c3