tast_tests: crostini: Increase timeout on RunWindowedApp
In tablet mode requesting a keyboard device creates a one off virtual
keyboard, which waits 5 seconds to give Chrome a chance to notice the
new device. This causes the crostini copy paste tests fail
consistently in tablet mode because we only give RunWindowedApp a 5
second timeout. Fix this by getting the keyboard device in the
crostini preconditions, so we don't have to repeatedly pay the time
cost during tests.

BUG=chromium:997564
TEST=tast run  crostini.CopyPasteArtifact*

Change-Id: Ib16f03806a6c278e093ef72e74e703dd36bae026