tast-test: getTaskInfo() works on N.
This CL adds support for activity.getTaskInfo() for ARC++ NYC.
"getTaskInfo()" is moved to the dumpsys.go file, and the
function is now called "DumpsysActivityActivities()".

DumpsysActivityActivities() checks the Android version.
If it is on NYC, it uses a regexp tailored for NYC (notice that NYC doesn't
support protobuf).
If it is Pie or higher, it uses the existing regexp. In a future CL, this
regexp will be replaced with the protobuf version.

It also makes taskInfo public (now TaskInfo). And the member variables
that are known be present on Protobuf are also made public.

The helper function, getBounds() was moved from activity.go to
dumpsys.go as well, where it makes more to sense to be.

TEST=Tests on P. Tested with arc.PIP, arc.ResizeActivity, works Ok.
     Tests on N. Tested with helper function that dumps the output of
     DumpsysActivityActivities and works ok.
BUG=chromium:1004471
Change-Id: I49f3a735d46521a3073b1b43930a42fa702dea02
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/1807997
Tested-by: Ricardo Quesada <ricardoq@chromium.org>
Auto-Submit: Ricardo Quesada <ricardoq@chromium.org>
Commit-Queue: Ricardo Quesada <ricardoq@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Stefan Kuhne <skuhne@chromium.org>