arc: Speculative fix for arc.PlayStorePersistent.
Based on test results, "Failed to wait daily hygiene is done" appears
on low-end devices. High-end devices do not have this error. This fix
increases the waiting timeout. Unfortunetly there is no logcat files
available on this failure, see example:
https://stainless.corp.google.com/browse/chromeos-autotest-results/swarming-4757fb6c04fcc110/
So fix is speculative. Also this dumps finsky shared prefs file on this
failure, however, it might not appear on bot results due to same reason
as logcat missing.

TEST=Locally, emulate this error
BUG=None

Change-Id: I9d13f485f65db6fe8937a89c41e2bb778f09c4bc