tast: fix ui.AssistantTimeQuery test failed.
We observed random time query test failure with error:
"Assistant time response doesn't match the current time",
when comparing Assistant time 2019-09-15 22:25:00 +0000 UTC with the
current time 2019-09-15 22:24:00 +0000 UTC. This is because during
comparison we did not include the cases where Assistant time equals
to the earliest or the latest timestamp of our acceptable time range.
This CL fix the comparison logic by including the above edge cases.

Some context:
In this test, Assistant time response may not exactly match with the
system now time due to possible clock drift and response latency.
Considering of this, we set a 1 min error tolerance and consider
Assistant gives the correct time response if it doesn't fall beyond
that range.

BUG=None.
TEST=tast -verbose run $(DUT) ui.AssistantTimeQuery

Change-Id: I32c365cff360735481aae746144e064729d43321