Update patch set 3
Patch Set 3:

(9 comments)

I think we should use HWDeps for display connection.

However, it is not yet ready, so meanwhile, display_backlight could be alternative. If it does not satisfy graphics team, runtime check with concrete TODO to switch to HWDeps when ready is the only workaround we can do for now. In the case, it's better to be written towards to the HWDeps direction a bit more, so commented inline.

Defer final call to nya@, as primary tast-reviewer.

Patch-set: 3
Reviewer: Gerrit User 1115919 <1115919@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>