Update patch set 1
Patch Set 1:

this channel stuff is a little out-of-the-ordinary. It makes more sense to me to do this with a lambda that returns an error (or nil).

I would push for the lambda approach as it is more amenable to e.g. providing arguments (like a handle to the running cmd, or its window(s)) whereas the pre-initialized channels dont have access to that info yet.

Can you do those things (easily) with channels? or, are there other channel advantages here?

Patch-set: 1
Reviewer: Gerrit User 1324920 <1324920@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>