Update patch set 1
Patch Set 1:

Wanna tag https://crbug.com/992469 in BUG= too?

Also, weren't there failures seen due to open fd(s) to /proc? I suppose that is probably a bug, not a feature, so you'd prefer to weed those out on the Chrome side? Or do you need to add to the whitelist?

Patch-set: 1
CC: Gerrit User 1132414 <1132414@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>