Update patch set 1
Patch Set 1:

Note: I got thrown off track because the parent CL is getting totally rewritten here. It would make more sense to me to squash the 2 CLs.

+ Alex, since he has made the same question on prior reviews

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > I did not review or discuss the original CL authored by Billy.
> > I am confused because I thought if shill is told to blacklist an interface, it wont include that interface when asked for the list of interfaces - I'll verify this now.

Via CLI or via something like org.chromium.flimflam.Manager.DisableTechnology? For the latter, the Device remains listed. For the former, it's skipped.

> > If its true, crawling sysfs still will be the correct option?

I'm not sure that follows. If the device is purposely blacklisted for some reason, then why not ignore it? And then if there are no other non-blacklisted devices...tests that expected to check something WiFi-related should probably fail, no?

> On a related note, (since Dean asked about replacing the sysfs crawling function with 
> querying shill):
> 1. Is there a reason autotest still relies on using sysfs.
> 2. We shouldn't need to rely on shill being running, to run a test like network_WlanDriver - querying sysfs is nice in that case.

On the other hand, we shouldn't expect Shill to be crashed/stopped/technology-blacklisted often, if we can avoid it. I think it is an antipattern to write too many tests that circumvent Shill. We've seen this on end-to-end WiFi tests (e.g., wifi_matfunc and similar) where we can't reliably screw around with 'iw ${DEV} scan' while shill/wpa_supplicant are also initiating scans -- and the answer shouldn't be to blacklist WiFi in shill, it should be to ask shill to ask wpa_supplicant to initiate scans on our behalf.

Regarding network_WlanDriver as a specific example: it used to ask Shill for the interface I think, but some silly details meant that certain odd devices were blacklisting WiFi? The right answer is to avoid those weird devices though (think "lab utility boxes"), not to bend over backwards to avoid Shill. In fact, I think looking back, skipping Shill was more of a speculative change, and I think it wasn't even needed (because we weren't actually running our network suites on the odd board variants).

But I also think that having generic "interface" routines ask Shill for answers is wrong: if we want to ask Shill, put the method in the 'shill' package. That will also make its limitations clearer.

So, if we really need both methods (and right now: it's not clear we do), there's enough room in the namespace for both -- network.WiFiInterface() and shill.WiFiInterface(). But we might as well just start with shill.WiFiInterface() (or similar -- I'm not good at naming).

Patch-set: 1