Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Dean confirmed that if shill is asked to blacklist a device, it wont show up in the list of devices returned by shill.
> 
> Why is that a problem? When do you expect to blacklist a WiFi device and still run tests against it?

The two actions (blacklisting a device, running a test against it) wont happen in the same
test. You provided an example of that in the commit message: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/481439

If you still think shill is the right place to query, I'd request that you formalize a dbus API to do so. I looked around in shill/dbus_bindings/ and don't see  GetInterfaceList or similar dbus call in the Manager's dbus-xml file or otherwise. 

Depending on shill to give you an accurate interface list without even formalizing the API (not that adding a dbus method to the interface is "formalizing" it, but its a step..)
seems wraught with danger to me. 


> 
> > With that and the other arguments above, this is a -1 from me unless others have objections.

Patch-set: 2
Reviewer: Gerrit User 1158909 <1158909@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>