tast.video: Add missing android dependancy to VideoDecoder tests.
The VideoDecodeAccelerator-based video decoders are currently being migrated to
the media::VideoDecoder interface. These video decoders are only supported on
platforms that support import mode. As we don't have a SoftwareDeps to only run
VD-specific tests on platforms supporting import mode, the "android"
SoftwareDeps is used, which guarantees a device supports import mode.

Unfortunately this dependancy was missing on multiple tests, causing devices
without import-mode support to run VD-specific tests. This CL adds the "android"
SoftwareDeps to the following tests:
- DecodeAccelVDH264ResolutionSwitch
- DecodeAccelVDVP8ResolutionSwitch
- DecodeAccelVDVP9ResolutionSwitch
- PlayVDH264
- PlayVDVP8
- PlayVDVP9
- PlayDecodeAccelUsedVDH264
- PlayDecodeAccelUsedVDVP8
- PlayDecodeAccelUsedVDVP9
- PlaybackVDPerfH2641080P60FPS
- PlaybackVDPerfVP81080P60FPS
- PlaybackVDPerfVP91080P60FPS
- SeekVDH264
- SeekVDSwitchH264

TEST=tests listed above on eve

BUG=b:141522148

Change-Id: I22af1a5f66de278e677effcaf69de7d0f100eda4