Wait for window state change before running tests in arc.PIP
As ActivityRecord#idle doens't work to wait for PIP transition to
be completed, we are currently using sleep(), which is not good.

As, with defer commit, now everything (window state, bounds, etc) is
committed at the same time, we can just wait for window state change
to ensure everything is ready.

This is also necessary to make sure the Chrome-side bounds is set
properly.

BUG=b:135956156
BUG=b:131248000
BUG=b:141588151
TEST=tast -verbose run my_crbook arc.PIP

Change-Id: I856dd2e31324712a5046b3edb80f3aef990f1706