Take off by one into account in testPIPToggleTabletMode
In arc.PIP, pipPositionErrorMarginPX is defined to handle various
rounding issues. We need to apply this to this test as well.

Also, we need to add sleep() to catch the problem. We are thinking
of adding a new sync logic, so once it's done, we can remove this
sleep().

BUG=b:135956156
BUG=b:141589435
TEST=tast -verbose run my_crbook arc.PIP

Change-Id: I6929ad34a9530b492e39d9f1e75f71f28ad4707f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/1824789
Reviewed-by: Stefan Kuhne <skuhne@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Kazuki Takise <takise@chromium.org>
Tested-by: Kazuki Takise <takise@chromium.org>
Auto-Submit: Kazuki Takise <takise@chromium.org>