Update patch set 1
Patch Set 1: Code-Review+1

lgtm from PIP perspective. 

The sleep looks suspicious, but this isn't an isolated problem with PIP - there are two other tests which have this problem theoretically as well, and it happens whenever you try to verify that nothing happens.

We have a solution for this but it's very complicated and I think it's a judgement call as to whether it's worth it compared to just adding a sleep.

I've checked all the usages of WaitForIdle in the ARC Tast tests, and there are currently only three instances in total of this type of problem.

Patch-set: 1
Reviewer: Gerrit User 1339199 <1339199@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Reviewer: Gerrit User 1179631 <1179631@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1