tast: use Fake Webcam for webrtc.PeerConn{H264,VP8}
This CL creates a specific precondition ChromeVideoWithFakeWebcam
for running WebRTC PeerConnection tests with a fake video/audio
capture device (a.k.a. "Fake Webcam" or "rolling pacman"). It also
uses this precondition for webrtc.PeerConn{H264,VP8}.

This CL also takes the opportunity to adjust owners of said tests.

This CL also reflows text in pre.go to fit in 80 columns, which
makes it more legible, albeit I think this is not enforced.

BUG=chromium:1008080
TEST=running webrtc.PeerConn{H264,VP8} on nocturne

Change-Id: Ib39fa1af1dc3e2cd0a0f50e87f02c39cba0cae22