tast: use Fake Webcam for webrtc.PeerConn{H264,VP8}Perf
This CL creates a specific precondition ChromeFakeCameraPerf for
running WebRTC PeerConnection Perf tests with a fake video/audio
capture device (a.k.a. "Fake Webcam" or "rolling pacman"). It
also uses this precondition for webrtc.PeerConn{H264,VP8}Perf.

This CL also fixes the comments on the aforementioned tests, and
takes the opportunity to adjust owners of said tests.

BUG=chromium:1008080
TEST=running webrtc.PeerConn{H264,VP8}Perf on nocturne

Cq-Depend: chromium:1825322
Change-Id: I0ce4a82e8fc03b3f6e333aa83d3912e910a0f9e8