Update patch set 1
Patch Set 1:

> 1. Can we move all of the inner functions outside of the main test function for readability?

Inner functions are actually the preferred way to write functions that are likely to only appear in this test.

That being said, the first function in the test body (which is defined anonymously and called immediately, and seems to use the function mostly to use return as control flow) can probably be factored out somewhere.

Patch-set: 1
Reviewer: Gerrit User 1117055 <1117055@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>