video.Play: Fix Tast video VDA tests using VD in some cases.
Recently the behavior of the ChromeosVideoDecoder feature was changed. This
feature controls enabling/disabling the new media::VideoDecoders, and is now
enabled by default on a percentage of all devices. But the logic
enabling/disabling the feature in the Tast tests still assumes that the feature
is disabled by default. This causes the tast.video.Playback and tast.video.Play
tests to incorrectly use the VD instead of VDA in some cases.

TEST=video.Playback(VD)PerfH2641080P60FPS and video.video.Play(VD)H264 on eve

BUG=chromium:141652665

Change-Id: I6e0f010cd1d2ae81c30a5e6241e6c5efdcca6524
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/1824562
Commit-Queue: David Staessens <dstaessens@chromium.org>
Tested-by: David Staessens <dstaessens@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>