tast-tests: crostini: Added recovery code to the precondition
Crostini tests would exhibit cascading failures when a test earlier on,
using the same precondition, trashed the vm/container state. We add a
simple mechanism to check that we can at least contact the running
container, and if we can't we reset the precondition and run it
again.

BUG=chromium:1001913
TEST=tast -verbose run --build $DUT crostini.Crash crostini.Launcher*

Change-Id: I3f80fdb8d185552565ccf76b47ce8f976388af23
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/1824566
Reviewed-by: Fergus Dall <sidereal@google.com>
Reviewed-by: Keigo Oka <oka@chromium.org>
Tested-by: Nic Hollingum <hollingum@google.com>
Commit-Queue: Nic Hollingum <hollingum@google.com>