Update patch set 1
Patch Set 1:

(1 comment)

Seems pretty reasonable to me.

It does look like we are doing this preamble where we set up the startup and cleanup contexts and then use them for setting up and tearing down the sludge VM and Wilco support daemon in a few places now, though. Is this going to show up in every Wilco test? Is it worth trying to find a way to factor out this setup and cleanup?

Patch-set: 1
Reviewer: Gerrit User 1117055 <1117055@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>