Update patch set 1
Patch Set 1:

(1 comment)

> Patch Set 1:
> 
> (1 comment)
> 
> Seems pretty reasonable to me.
> 
> It does look like we are doing this preamble where we set up the startup and cleanup contexts and then use them for setting up and tearing down the sludge VM and Wilco support daemon in a few places now, though. Is this going to show up in every Wilco test? Is it worth trying to find a way to factor out this setup and cleanup?

It should just show up in these three `sludge` tests. I thought about how this could be better condensed, but from my perspective it seems as abstracted as it should be. It would be a little weird of the library functions had internal context timeouts that were not obvious to the user. Also through the three tests, sometimes we are starting only the VM, and sometimes we need both the VM and the helper daemon. If you see any clean way to refactor it, let me know.

Patch-set: 1
Label: Commit-Queue=0