Update patch set 8
Patch Set 8:

> Patch Set 8:
> 
> David, any update on this?
> https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/1741728/2//COMMIT_MSG
> 
> Please let me know when I can review this again.

Most issues with the new video decoder tests are fixed, but there is still a crash that happens on rare occasions. I haven't been able to reproduce this, so I'm adding code to collect crash dumps. I'm also creating a CL to enabled test-related DCHECKS in release, to hopefully get a clue as to why it crashes.

There is also an issue that causes frame validation to fail from time to time on some platforms.

I think it will be very hard to get a 100% stability in the near future, and the issues we're detecting might actually be real issues, as the new tests perform a lot more checks. So I guess the best thing to do is submit this CL as it's blocking all my other VDA-test-removal related CLs. It's still pretty quick and easy to detect bugs by using our alert mails and the bisector.

WDYT?

Patch-set: 8
Reviewer: Gerrit User 1288602 <1288602@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>