security.SELinux: Avoid races on recursive scans.
When recursive scanning files, a file can be added or removed in
race, especially on special file systems like sysfs and procfs.

To avoid such races fundamentally, ignore system call errors
(lstat and readdir) when recursively scanning files.

By default, errors about the topmost file of recursive scans are
not ignored. CheckContextReq.IgnoreErrors can be set to true to
ignore these errors too. This is useful when the topmost file
might not exist.

selinux.SkipNotExist is replaced with IgnoreErrors. As a bonus,
we can guarantee that os.FileInfo passed to FileLabelCheckFilter
is always non-nil.

BUG=chromium:1009309
TEST=fast_build.sh -T
TEST=tast run betty security.SELinuxFilesSystem

Change-Id: I85ce643835e7801bfd8165e24a748b70e5c39600