tast: always check the % of broken frames for PeerConn
This CL splits checkVideoHealth() into checkTotalFrames() and
checkBrokenFrames(); this allows RunPeerConn() to always
check for broken frames, leaving RunGetUserMedia() functionally
unchanged.

RunPeerConn() should use a Fake webcam and in this case we want
to always check for broken frames, RunGetUserMedia() should only
check for for broken frames when running in a VM (e.g. QEMU).

BUG=chromium:1008080
TEST=running webrtc.PeerConnVP8 and camera.GetUserMedia on nocturne

Change-Id: I39521cec45c2cbd2a342c58e833fb4aae552b1c3